requirements.txt
Flask==3.0.3
gunicorn==22.0.0
torch==2.3.0
torchvision==0.18.0
Pillow==10.3.0
albumentations==1.3.1
opencv-python-headless==4.9.0.80
